+++ app/src/main/res/drawable/custom_button.xml
@@ -1,5 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
-    <gradient android:startColor="#2AF598" android:endColor="#009EFD" android:type="linear"/>
+    <gradient android:startColor="#00cec9" android:endColor="#00cec9" android:type="linear"/>
+    <stroke
+        android:width="2dp"
+        android:color="#fff"/>
+
     <corners android:radius="20dp"/>
+
 </shape>
