+++ app/src/main/res/layout/activity_restaurant_details.xml
@@ -156,8 +156,8 @@
         android:id="@+id/restTimings"
         android:layout_width="335dp"
         android:layout_height="47dp"
-        android:fontFamily="@font/segoeui"
         android:text="timings"
+        android:fontFamily="@font/segoeui"
         android:textColor="#FFFFFF"
         app:autoSizeMaxTextSize="20sp"
         app:autoSizeMinTextSize="15sp"
